唉内存溢出，现在好了，能正常跑二周目了

+++ User/adc/adcmod.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 void ADCMod_Init(void);
-double ADCMod_GetKnob(void);
+uint16_t ADCMod_GetKnob(void);
 uint32_t ADCMod_GetRandom(void);
 
 #endif
